refactor: 중복 인증 응답 생성 로직을 AuthenticationResponseUtil로 분리 및 개선
