+++ src/main/java/com/rabbitmqprac/application/api/UserApi.java
@@ -1,6 +1,6 @@
 package com.rabbitmqprac.application.api;
 
-import com.rabbitmqprac.application.dto.auth.res.UserDetailRes;
+import com.rabbitmqprac.application.dto.user.res.UserDetailRes;
 import com.rabbitmqprac.application.dto.user.req.NicknameCheckReq;
 import com.rabbitmqprac.application.dto.user.req.NicknameUpdateReq;
 import com.rabbitmqprac.domain.context.user.exception.UserErrorCode;
@@ -8,7 +8,6 @@ import com.rabbitmqprac.global.annotation.ApiExceptionExplanation;
 import com.rabbitmqprac.infra.security.authentication.SecurityUserDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.RequestBody;
