+++ src/main/java/com/rabbitmqprac/application/controller/AuthController.java
@@ -4,16 +4,13 @@ import com.rabbitmqprac.application.api.AuthApi;
 import com.rabbitmqprac.application.dto.auth.req.AuthSignInReq;
 import com.rabbitmqprac.application.dto.auth.req.AuthSignUpReq;
 import com.rabbitmqprac.application.dto.auth.req.AuthUpdatePasswordReq;
+import com.rabbitmqprac.application.dto.user.res.UserIdRes;
 import com.rabbitmqprac.domain.context.auth.service.AuthService;
-import com.rabbitmqprac.global.util.CookieUtil;
+import com.rabbitmqprac.global.util.AuthenticationResponseUtil;
 import com.rabbitmqprac.infra.security.authentication.SecurityUserDetails;
-import com.rabbitmqprac.infra.security.jwt.Jwts;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.tuple.Pair;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
@@ -24,9 +21,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.time.Duration;
-import java.util.Map;
-
 @Slf4j
 @RequiredArgsConstructor
 @RequestMapping("/auth")
@@ -36,14 +30,14 @@ public class AuthController implements AuthApi {
 
     @Override
     @PostMapping("/sign-up")
-    public ResponseEntity<Map<String, Long>> signUp(@RequestBody @Validated AuthSignUpReq authSignUpReq) {
-        return createAuthenticatedResponse(authService.signUp(authSignUpReq));
+    public ResponseEntity<UserIdRes> signUp(@RequestBody @Validated AuthSignUpReq authSignUpReq) {
+        return AuthenticationResponseUtil.createAuthenticatedResponse(authService.signUp(authSignUpReq));
     }
 
     @Override
     @PostMapping("/sign-in")
-    public ResponseEntity<Map<String, Long>> signIn(@RequestBody @Validated AuthSignInReq authSignInReq) {
-        return createAuthenticatedResponse(authService.signIn(authSignInReq));
+    public ResponseEntity<UserIdRes> signIn(@RequestBody @Validated AuthSignInReq authSignInReq) {
+        return AuthenticationResponseUtil.createAuthenticatedResponse(authService.signIn(authSignInReq));
     }
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
@@ -53,17 +47,5 @@ public class AuthController implements AuthApi {
         authService.updatePassword(user.getUserId(), authUpdatePasswordReq);
     }
 
-    private ResponseEntity<Map<String, Long>> createAuthenticatedResponse(Pair<Long, Jwts> userInfo) {
-        ResponseCookie cookie = CookieUtil.createCookie(
-                "refreshToken", userInfo.getValue().refreshToken(), Duration.ofDays(7).toSeconds()
-        );
-
-        return ResponseEntity.ok()
-                .header(HttpHeaders.SET_COOKIE, cookie.toString())
-                .header(HttpHeaders.AUTHORIZATION, userInfo.getValue().accessToken())
-                .body(
-                        Map.of("userId", userInfo.getKey())
-                );
-    }
 
 }
